this was an effort to move from caching files in individual files to caching a one big file
well, not one big file:
- using N partitions as big files, with the main goal of being able to compact one partition at-a-time
  this limits the overhead of partitioning to 1/(N+1)
- having this setup on two levels, (i) for small ranges and (ii) big ranges
  the idea would be that (i) small ranges (e.g. 512b) go to a different bigfile than large ranges
  (ii) there is indepedent LRU for small and large ranges and in principles 10% of
  capacity is for small ranges, which means that when the cache is full and ranges get evicted
  this will only affect large ranges. This is beneficial because HTTP transfers of small ranges
  are limited by latency making them much more expensive per byte so one should prefer these small
  ranges in the caceh. Note that for parquet, the small ranges are the file footer, and rowgroup
  headers, which get read often (this policy holds unless the large file is smaller than 90% of target
  in which case the rest capacity is given to small ranges -- this allows workloads that exclusive
  consist of small ranges such as a directory full of small json or xml files to use the whole cache).

but.. I got a bit stuck on tackling the multi-threaded aspect of this. I want the background writers
to also handle compaction. I was heading for a redesign where we use on thread per partition as
this would removes conflicts due to two threads writing into the same file, and also naturally
pauses additional writing to a partition that is being compacted.

here, I ran out of confidence claude could do this for me automatically and the current state is
still noncompiling and nonworking.

I am shelving this approach for now because the main motivation to go to one bigfile (or few) was
that eventually I wanted to create later a sideways entrance into the cache using io_uring
and io_uring benefits if the files are already open (which means the amount of files must be low,
here I was working with 16 partitions, so 32 files: one for small and one for large ranges).

But, new thinking about hybrid indexing in DuckDB for DuckLake made me consider, moving the PDX
vector index back into DuckDB storage in its block manager. It is also an acknowledgement that maybe
going for an approach with a cold cache and many random I/Os will lose in many instances to loading
a partition with (one or more) 256KB reads, as DuckDB does - in which case the benefit from io_uring
over even sync I/O is not shocking (and it can be better than doing many small accesses).
The approach that integrates PDX as a proper DuckDB index has also other advantages, in handling
transactions. So with that motivation of io_uring and few files reduced, I decided to stall
this branch for now.