The extension now uses DuckDB's ExternalFileCache on the internal cache-files of BlobCache to get swappable RAM caching through the DuckDB buffer manager.
We renamed our structs to start with BlobCache since the ExternalFileCache
has some classes with similar names.

The extension now automatically does set external_file_cache = false
so the parquet reader will not use the normal ExternalFileCache in the dbinstance.
Rather, its reads will be cached by blobcache, on disk and in RAM.