renamed is CacheFile into CacheEntry.
Introduced a layer of indirection below CacheRange that is the CacheFile. The LRU of a CacheMap should no longer be on CacheEntry but on CacheFile records. The data belonging to multiple CacheRanges will be stored in each CacheFile.

Eviction hence targets CacheFiles. When a CacheFile gets deleted it deletes all CacheRange that refer to it. But this does not happen immediately but lazily: when we look up a CacheRange via AnalyzeRange and we if find one without a CacheFile we delete it lazily from the map of CacheRanges and move to the next (possibly also deleting it etc).

CacheFiles are kept in a map with as key the filepath of the CacheFile and as value a unique_ptr to the CacheFile. The map is stored in the CacheMap. It thus now will contain a map from original files (cache_key) to CacheEntry and one between filepath and CacheFile. The LRU mechanism targets the latter.

The small range git incremented to 16KB. The largerange cache now uses a policy of storing each of its CacheRanges in a separate CacheFile. The effect is that the large CacheFiles are at least 16KB. Small CacheFiles use a different policy: they contain multiple ranges and grow over time to <256KB. That is, if the addition of one more CacheRange would make the smallrange CacheFile grow over 256KB and we start appending to a new CacheFile. In the memcache, we always look up ranges with offset==0. The memcpy happens from the appropriate offset of the range in that file. This limits memory fragmentation in DuckDB.

In the case of small ranges we append the first small access to a datafile in CacheFiles named “Xfirst” in a subdirectory hashed by X and all others smallranges  in files named “Xsmall”. These are X numbered files starting at 10000000. This allows us to potentially have many of these files without clogging one directory. A new file is started with an incremented number when the previous would exceed 256KB.

In te naming scheme for largerange CacheFiles we remove the s/l prefix and add instead the range start as a number. We still need to use the id because the same or overlapping range can get cached, evicted and cached again while we do not want these to use the same filename (because when a read reads an evicted file it should just fail and fall back to the original file, rather than hitting some newer file than it intended to read).

When a file is evicted we simply delete it from disk and from the CacheFile kvmap.

Rationale

The goal of this rewrite is to make the LRU for large ranges no longer based on the original filename only. This will make caching of parquet files more finegrained. Because, if the query pattern in one file changed from say select max(A) from t to max(B) we would not get rid of the A ranges. But now, they are in separate CacheFiles with separate LRU ranks and the cold ones will get evicted.

Another goal is to use in memcache for the small ranges pages of 256KB than contain many ranges. Previously the ranges could be as small as 256bytes or less and this means that if the smallrange cache contains a lot of data, there are a huge amount of DuckDB buffer memory ranges that will make eviction very cumbersome and may introduce a lot of allocation space and cpu overhead. Note that the policy of caching the first small access to a file in one special file targets the fact that parquet readers often read the footer (the last bytes of the parquet). We want to keep these around even longer and even hotter in the memcache than other small ranges.